# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Adpoint API",
        "contact": {
            "email": "alkhamilnaz@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        },
        {
            "url": "https://127.0.0.1:8000/api",
            "description": "Laravel Adpoint API"
        }
    ],
    "paths": {
...
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Display a listing of the categories",
                "operationId": "list categories",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/add": {
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Add Category",
                "operationId": "Add Category",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategorySchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Display a listing of the category by id",
                "operationId": "list category by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of category to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete category by id",
                "operationId": "Delete category by id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/category/edit/{id}": {
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Edit of the category by id",
                "operationId": "Edit category",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of edit category",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CategorySchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "Login user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "nazmudin@imaniprima.com",
                                    "password": "nazmudin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    },
                    "401": {
                        "description": "unauthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CategorySchema": {
                "title": "Category",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 0
                    },
                    "name": {
                        "type": "string",
                        "example": "Videotron"
                    },
                    "banner": {
                        "type": "string",
                        "example": "uploads/icon.jpg"
                    },
                    "icon": {
                        "type": "string",
                        "example": "uploads/banner.jpg"
                    },
                    "slug": {
                        "type": "string",
                        "example": "Videotron"
                    },
                    "meta_title": {
                        "type": "string",
                        "example": "Videotron"
                    },
                    "meta_description": {
                        "type": "string",
                        "example": "Videotron"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}